create CustomerCenterScreen
