+++ examples/paywall-tester/src/main/java/com/revenuecat/paywallstester/ui/screens/AppScreen.kt
@@ -4,6 +4,6 @@ sealed class AppScreen(val route: String) {
     object Main : AppScreen("main")
     object Paywall : AppScreen("paywall")
     object PaywallFooter : AppScreen("paywall_footer")
-
     object PaywallByPlacement : AppScreen("paywall_by_placement")
+    object CustomerCenter : AppScreen("customer_center")
 }
